requirements.txt
ckantoolkit>=0.0.7
requests>=2.28.1
